pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "email-validator>=2.2.0",
    "flask[async]>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pillow>=11.1.0",
    "playwright>=1.50.0",
    "psycopg2-binary>=2.9.10",
    "scraper>=0.1.0",
    "trafilatura>=2.0.0",
    "python-multipart>=0.0.20",
    "jinja2>=3.1.5",
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "typing-extensions>=4.12.2",
    "hypercorn>=0.17.3",
    "quart>=0.20.0",
    "requests>=2.32.3",
    "beautifulsoup4>=4.13.3",
]
